Fix page navigation after site redesign
The site we're scraping has been redesigned, so the HTML we're looking
for doesn't exist anymore. Due to this, the scraper has been crashing
for a while, at the point where we try to build our `page_urls` to
figure out what pages we're going to scrape. That's fine, we can fix
that!

The pagination links have changed - instead of having all of the
pagination links available on the first page, the list of links is now
truncated in the middle. e.g.

```
1 2 … 4 Next »
```

That means we can't access all of the links up-front, so I figure our
best bet is to scrape one page at a time, then look to see if there's a
"Next" link on that page to see if there's another page to scrape next.
Keep going until we get to a page without a "Next" link, and then we
know we're done! Easy!

This does that. The scraper is now running and looking at each page in
sequence. The `get_products_from_page` method is not finding any
products though, because the markup for the products has changed too.
Gotta fix that next!