Fix product building after site redesign
The markup for each product has changed, so this updates the logic that
extracts info from the markup.

The URL for each product has also changed, and it's the URL that we use
to determine if the product is unique. That means all existing products
will be counted as new products, so I should probably wipe all the
existing products from the database before deploying this.