Move some more RRR logic into the RRR module
Whoops, this logic that figures out which of the scraped products are
available will be specific to this one site, so it should live in this
module. Now the module's `.get_products` method returns only the
available products.

I've also added an initial `puts` that prints out which site we're
scraping. I'll do the same for the next site I add.